@@ -13,10 +13,10 @@
   "private": false,
   "sideEffects": false,
   "scripts": {
-    "prebuild": "rimraf dist && rimraf dist-modules-temp && node ./scripts/prebuild.js",
+    "prebuild": "rimraf ../../dev-plugins/components && rimraf ../../dev-plugins/components-modules-temp && node ./scripts/prebuild.js",
     "build": "npm run build:common && npm run build:modules",
-    "build:common": "BABEL_ENV=production babel --root-mode upward src --out-dir dist",
-    "build:modules": "BABEL_ENV=production BABEL_MODULES=1 babel --root-mode upward src --out-dir dist-modules-temp",
+    "build:common": "BABEL_ENV=production babel --root-mode upward src --out-dir ../../dev-plugins/components",
+    "build:modules": "BABEL_ENV=production BABEL_MODULES=1 babel --root-mode upward src --out-dir ../../dev-plugins/components-modules-temp",
     "postbuild": "BABEL_ENV=production babel-node --root-mode upward ./scripts/postbuild.js",
     "release": "npm run build && npm publish dist"
   },
@@ -57,7 +57,7 @@
     "babel-loader": "~7.1.2",
     "babel-plugin-module-resolver": "~3.1.3",
     "babel-plugin-rewire-exports": "~1.0.1",
-    "babel-plugin-styled-components": "~1.10.0",
+    "babel-plugin-styled-components": "^2.0.6",
     "chalk": "~1.1.3",
     "check-prop-types": "~1.1.2",
     "composable-form-tests": "~1.0.1",
@@ -93,8 +93,9 @@
   "peerDependencies": {
     "@reactioncommerce/components-context": "~1.2.0",
     "prop-types": "~15.7.2",
-    "react": "~16.9.0",
+    "react": "17.0.0",
     "react-container-query": "~0.11.0",
+    "react-dom": "17.0.0",
     "react-stripe-elements": "~2.0.1",
     "reacto-form": "~1.4.0",
     "styled-components": "~3.3.3"
