@@ -349,7 +349,7 @@ class TextInput extends Component {
     };
   }
 
-  componentWillMount() {
+  componentDidMount() {
     const { value } = this.state;
     this.handleChanging(value);
     this.handleChanged(value);
